:truck: rename: BaseEntity 클래스를 global 패키지 하위로 이동
