+++ src/main/java/com/nexters/palang/domain/book/domain/Book.java
@@ -1,6 +1,6 @@
 package com.nexters.palang.domain.book.domain;
 
-import com.nexters.palang.domain.common.BaseEntity;
+import com.nexters.palang.global.common.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
